joplen/requirements.txt
matplotlib==3.8.0
numpy==1.24.4
scikit-learn>=1.2.0
tqdm==4.65.0
cupy-cuda12x==12.2.0
xgboost==2.0.1
linear-tree==0.3.5
